requirements.txt
# 车牌识别系统依赖包

# 核心图像处理
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
scikit-image>=0.21.0
imutils>=0.5.4

# OCR引擎
pytesseract>=0.3.10
paddlepaddle>=2.5.0
paddleocr>=2.7.0
easyocr>=1.7.0

# 深度学习框架
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.15.0

# Google AI
google-generativeai>=0.3.2

# Web框架
flask>=2.3.0
flask-cors>=4.0.0
flask-restx>=1.1.0
werkzeug>=2.3.0

# API和HTTP
requests>=2.31.0
urllib3>=2.0.0
gunicorn>=21.0.0

# 数据分析和可视化
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# 机器学习
scikit-learn>=1.3.0
scipy>=1.11.0
joblib>=1.3.0

# 图像增强
albumentations>=1.3.0
imgaug>=0.4.0

# 配置和序列化
pyyaml>=6.0.1
toml>=0.10.2
configparser>=5.3.0

# 命令行工具
click>=8.1.0
argparse>=1.4.0
colorama>=0.4.6
rich>=13.0.0

# 系统监控和性能
psutil>=5.9.0
prometheus-client>=0.17.0
memory-profiler>=0.61.0
py-spy>=0.3.14

# 日志和调试
loguru>=0.7.0
colorlog>=6.7.0
structlog>=23.0.0

# 进度条和时间
tqdm>=4.65.0
python-dateutil>=2.8.2

# 文件处理和路径
pathlib2>=2.3.7
watchdog>=3.0.0

# 并发和异步
concurrent-futures>=3.1.1
asyncio>=3.4.3

# 测试框架
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-benchmark>=4.0.0

# 数据格式
jsonschema>=4.18.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# 类型检查和静态分析
mypy>=1.5.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# 安全和加密
cryptography>=41.0.0
bcrypt>=4.0.0

# 缓存
redis>=4.6.0
diskcache>=5.6.0

# 其他工具
python-dotenv>=1.0.0
humanize>=4.7.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0